# level2/task2/docs/swagger.yaml
swagger: "2.0"
basePath: /api/v1
definitions:
  controller.AccountBalance:
    properties:
      address:
        example: 0x742d35Cc6634C0532925a3b844Bc454e4438f44e
        type: string
      balance:
        example: "1000000000000000000"
        type: string
    type: object
  controller.AccountResponse:
    properties:
      address:
        example: 0x742d35Cc6634C0532925a3b844Bc454e4438f44e
        type: string
      balance:
        example: "1000000000000000000"
        type: string
    type: object
  controller.ArrayDemoAddValueRequest:
    properties:
      value:
        example: "42"
        type: string
    required:
    - value
    type: object
  controller.ArrayDemoAddValueResponse:
    properties:
      address:
        type: string
      txHash:
        type: string
      value:
        type: string
    type: object
  controller.ArrayDemoGetValuesResponse:
    properties:
      address:
        type: string
      count:
        type: integer
      values:
        items:
          type: string
        type: array
    type: object
...
  controller.ErrorResponse:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: 服务器内部错误
        type: string
    type: object
...
  controller.TransactionResponse:
    properties:
      blockNumber:
        example: "12345"
        type: string
      from:
        example: 0x742d35Cc6634C0532925a3b844Bc454e4438f44e
        type: string
      hash:
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        type: string
      timestamp:
        example: 1634567890
        type: integer
      to:
        example: 0x742d35Cc6634C0532925a3b844Bc454e4438f44e
        type: string
      value:
        example: "1000000000000000000"
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 提供以太坊区块链相关的API服务，包括账户管理、交易处理、合约部署等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Web3 区块链接口服务
  version: "1.0"
paths:
  /accounts/{address}/balance:
    get:
      consumes:
      - application/json
      description: 获取指定以太坊地址的账户余额
      parameters:
      - description: 以太坊地址
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.AccountBalance'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 获取账户余额
      tags:
      - 账户
    put:
      consumes:
      - application/json
      description: 设置指定以太坊地址的账户余额
      parameters:
      - description: 以太坊地址
        in: path
        name: address
        required: true
        type: string
      - description: 余额请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.AccountBalance'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 设置账户余额
      tags:
      - 账户
  /accounts/{address}/code:
    get:
      consumes:
      - application/json
      description: 获取指定以太坊地址的合约代码，如果地址是合约地址则返回其代码
      parameters:
      - description: 以太坊地址
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 获取账户合约代码
      tags:
      - 账户
  /accounts/{address}/nonce:
    get:
      consumes:
      - application/json
      description: 获取指定以太坊地址的账户当前Nonce值
      parameters:
      - description: 以太坊地址
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 获取账户Nonce值
      tags:
      - 账户
  /accounts/{address}/transactions:
    get:
      consumes:
      - application/json
      description: 获取指定以太坊地址的交易历史记录
      parameters:
      - description: 以太坊地址
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.TransactionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 获取账户交易列表
      tags:
      - 账户
  /accounts/hd-wallet:
    post:
      consumes:
      - application/json
      description: 创建新的以太坊分层确定性钱包
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 创建 HD 钱包
      tags:
      - 账户
  /accounts/keystore:
    post:
      consumes:
      - application/json
      description: 创建新的以太坊账户密钥库文件
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.AccountResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 创建密钥库
      tags:
      - 账户
  /accounts/wallet:
    post:
      consumes:
      - application/json
      description: 创建新的以太坊账户钱包
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.AccountResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 创建钱包
      tags:
      - 账户
...
  /contract/arraydemo/add-value:
    post:
      consumes:
      - application/json
      description: 向数组合约添加新的整数值
      parameters:
      - description: 添加值请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ArrayDemoAddValueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ArrayDemoAddValueResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 向ArrayDemo合约添加值
      tags:
      - 合约操作
  /contract/arraydemo/get-values:
    get:
      consumes:
      - application/json
      description: 获取ArrayDemo合约的所有值
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ArrayDemoGetValuesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 获取ArrayDemo合约的值
      tags:
      - 合约操作
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey